Внечены изменения в учет материалов и в типы данных идентификаторов (ключи bigInteger в некоторых местах заменены на integer)

+++ database/migrations/2020_11_03_065622_create_q3w_materials_table.php
@@ -15,7 +15,7 @@ class CreateQ3wMaterialsTable extends Migration
     {
         Schema::create('q3w_materials', function (Blueprint $table) {
             $table->bigIncrements('id')->comment('Уникальный идентификатор');
-            $table->bigInteger('standard_id')->unsigned()->comment('Идентификатор эталона')->index();
+            $table->integer('standard_id')->unsigned()->comment('Идентификатор эталона')->index();
             $table->integer('project_object')->unsigned()->comment('Идентификатор объекта')->index();
             $table->integer('amount')->unsigned()->nullable()->comment('Количество в штуках (для штучного учета)');
             $table->double('quantity')->unsigned()->comment('Количество в единицах измерения');
